requirements.txt
arize-phoenix
arize-phoenix-otel
openinference-instrumentation-openai
openai